Move editor widget into main src/
